pyproject.toml
[tool.poetry]
name = "agrimarket-nepal"
version = "0.1.0"
description = "Final Year Project"
authors = ["CEPHAL0 <xaradxarma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.django-stubs]
django_settings_module =  "myproject.settings"
